\name{SiPSiC-package}
\alias{SiPSiC-package}
\alias{SiPSiC}
\docType{package}
\title{
\packageTitle{SiPSiC}
}
\description{
\packageDescription{SiPSiC}
}
\details{
Use this package to calculate per-cell scores for a biological pathway of your choice, from single-cell RNA-seq data. Transcripts-Per-Million (TPM) or Counts-Per-Million (CPM) units of the data are recommended for best results.
}
\author{
Daniel Davis, Yotam Drier. Maintainer: Daniel Davis
}
\references{
\url{https://medicine.ekmd.huji.ac.il/en/research/yotamd/Pages/default.aspx}
}
\seealso{
\code{\link{getPathwayScores}}
}